bunch of changes, primarily to newlib/libgloss
addresses compilation issues, changes some syscall struct details such as
stat, fakes some c-library headers, and adds a new "brgarm" backend with
custom crt1 etc.

+++ src/newlib-2.1.0/newlib/libc/include/sys/_types.h
@@ -17,14 +17,20 @@ typedef long _off_t;
 #endif
 
 #ifndef __dev_t_defined
-typedef short __dev_t;
+//typedef short __dev_t;
+// berkin hack:
+typedef long long __dev_t;
 #endif
 
 #ifndef __uid_t_defined
-typedef unsigned short __uid_t;
+//typedef unsigned short __uid_t;
+// berkin hack:
+typedef unsigned long __uid_t;
 #endif
 #ifndef __gid_t_defined
-typedef unsigned short __gid_t;
+//typedef unsigned short __gid_t;
+// berkin hack:
+typedef unsigned long __gid_t;
 #endif
 
 #ifndef __off64_t_defined
