+++ src/newlib-2.1.0/newlib/libc/include/sys/types.h
@@ -152,7 +152,9 @@ typedef	char *	caddr_t;
     defined(__sparc__) || defined(__SPU__)
 typedef	unsigned long	ino_t;
 #else
-typedef	unsigned short	ino_t;
+//typedef	unsigned short	ino_t;
+// berkin hack
+typedef	unsigned long	ino_t;
 #endif
 #endif /*__CYGWIN__*/
 
@@ -219,7 +221,9 @@ typedef unsigned int mode_t _ST_INT32;
 #endif /* ! __MS_types__ */
 #endif /*__CYGWIN__*/
 
-typedef unsigned short nlink_t;
+//typedef unsigned short nlink_t;
+// berkin hack:
+typedef long nlink_t;
 
 /* We don't define fd_set and friends if we are compiling POSIX
    source, or if we have included (or may include as indicated
